% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_data.R
\name{add_trivia_class}
\alias{add_trivia_class}
\title{add_trivia_class}
\usage{
add_trivia_class(parsed_data)
}
\arguments{
\item{parsed_data}{The output of `parse_raw_data`}
}
\value{
A modified dataframe (has classes added)
}
\description{
Inspects the dimensions of the returned data to determine whether the question is a multiple choice or a boolean question. This is only ever called internally
}
\details{
The API returns with a field for the correct answer, and then either 1 (boolean) or 3 (multiple choice) fields for the incorrect answers. Thus, you can determine what type of question it is by inspecting the dimensions.

Since the two types of questions have different fields, a different class is assigned for each type. This is then passed to a relevant s3 method to parse this data into a common form with a 'correct answer' field and an 'incorrect answers' field
}
