% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categories.R
\docType{data}
\name{categories}
\alias{categories}
\title{Categories data}
\format{
A dataset containing 24 rows and 2 variables
\describe{
\item{id}{a unique identifier for each category}
\item{categories}{name of the category}
}
}
\usage{
categories
}
\description{
A dataset all allowed categories
}
\examples{
data(categories)
}
\keyword{datasets}
