% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_categories.R
\name{get_categories}
\alias{get_categories}
\title{get_categories}
\usage{
get_categories()
}
\value{
A 2 column dataframe of category and id
}
\description{
Get all available categories and their ids
}
\examples{
library(opentriviadb)
get_categories()
}
