% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_questions.R
\name{get_questions}
\alias{get_questions}
\title{get_questions}
\usage{
get_questions(
  number = 10,
  category,
  difficulty = c("any", "easy", "medium", "hard"),
  type = c("any", "multiple", "boolean")
)
}
\arguments{
\item{type}{}
}
\value{

}
\description{
Document
}
