% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.R
\name{isConnected}
\alias{isConnected}
\title{Is the current session (networked) and connected?}
\usage{
isConnected(site = "https://opentdb.com/")
}
\arguments{
\item{site}{Character variable with site to try to connect to,
defauls to \code{http://www.opentdb.com}}
}
\value{
A boolean value indicating networking status
}
\description{
Function to (heuristically) test for a network connection by attempting
to connect to a given website.
}
\details{
This function was kindly provided by Dirk Eddelbuettel as part of this SO
answer (https://stackoverflow.com/questions/66773611/cran-example-fails-if-run-offline-talks-to-external-api). It is copied here to avoid the `dang` dependency.
}
\author{
Dirk Eddelbuettel
}
