% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_questions.R
\name{make_category_string}
\alias{make_category_string}
\title{make_category_string}
\usage{
make_category_string(category)
}
\arguments{
\item{category}{The trivia category}
}
\description{
Builds a string for the category block of the API
}
