% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opentriviadb-package.R
\docType{package}
\name{opentriviadb-package}
\alias{opentriviadb}
\alias{opentriviadb-package}
\title{opentriviadb: Interface to the 'Open Trivia Database'}
\description{
The 'Open Trivia Database'<https://opentdb.com/> is a open source,
    user contributed, database of trivia questions and answers. This package 
    provides an interface to download and access questions and answers from
    within R.
}
\keyword{internal}
