% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WIP.R
\name{parse_raw_data}
\alias{parse_raw_data}
\title{parse_raw_data}
\usage{
parse_raw_data(unparsed_data)
}
\arguments{
\item{unparsed_data}{The result of a `GET` call to the OpenTriviaDB API}
}
\value{
A dataframe of either 6 or 8 columns
}
\description{
Parses the initial response from the API call. Output can have either 6 columns (boolean question) or 8 columns (multiple choice question).
}
