% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_trivia_df.R
\name{parse_trivia_df.default}
\alias{parse_trivia_df.default}
\title{parse_trivia_df.default}
\usage{
\method{parse_trivia_df}{default}(data)
}
\arguments{
\item{data}{}
}
\value{

}
\description{
parse_trivia_df.default
}
