% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_trivia_df.R
\name{parse_trivia_df}
\alias{parse_trivia_df}
\title{parse_trivia_df}
\usage{
parse_trivia_df(data)
}
\arguments{
\item{data}{A dataframe with either the 'trivia_multi_choice' or 'trivia_boolean' class. Usually this is the output of the `add_trivia_class` function}
}
\value{
a list of dataframes
}
\description{
Generic function for turning either a boolean or multi-choice question into a dataframe with the 'possible_answers' field. Used internally by the {{}} function.
}
