% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_trivia_df.R
\name{parse_trivia_df.trivia_multi_choice}
\alias{parse_trivia_df.trivia_multi_choice}
\title{parse_trivia_df.trivia_multi_choice}
\usage{
\method{parse_trivia_df}{trivia_multi_choice}(data)
}
\arguments{
\item{data}{}
}
\value{

}
\description{
parse_trivia_df.trivia_multi_choice
}
